+++ domain/src/main/java/com/devlabs/domain/entity/Activity.kt
@@ -7,5 +7,5 @@ data class Activity(
     val price: Float,
     val link: String,
     val key: String,
-    val accessibility: Int
+    val accessibility: Float
 )
